Safari-611.1.10/LayoutTests/fast/text/simple-line-layout-text-stroke-width-expected.txt

PASS internals.repaintRectsAsText().indexOf('90 8 62 18') is not -1
PASS successfullyParsed is true

TEST COMPLETE
foobar
